// web/src/generated/api/models/GroupSummaryResponse.ts
export interface GroupSummaryResponse {
    /**
     * 
     * @type {number}
     * @memberof GroupSummaryResponse
     */
    expensesCount?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GroupSummaryResponse
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof GroupSummaryResponse
     */
    memberCount?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GroupSummaryResponse
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof GroupSummaryResponse
     */
    outingsCount?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GroupSummaryResponse
     */
    ownerAccountId: string;
    /**
     * 
     * @type {number}
     * @memberof GroupSummaryResponse
     */
    participantsCount?: number | null;
    /**
     * 
     * @type {string}
     * @memberof GroupSummaryResponse
     */
    role: GroupSummaryResponseRoleEnum;
    /**
     * 
     * @type {string}
     * @memberof GroupSummaryResponse
     */
    settlementPolicy: GroupSummaryResponseSettlementPolicyEnum;
}


/**
 * @export
 */
export const GroupSummaryResponseRoleEnum = {
    Owner: 'owner',
    Member: 'member'
} as const;
export type GroupSummaryResponseRoleEnum = typeof GroupSummaryResponseRoleEnum[keyof typeof GroupSummaryResponseRoleEnum];

/**
 * @export
 */
export const GroupSummaryResponseSettlementPolicyEnum = {
    OwnerOnly: 'owner_only',
    AnyMember: 'any_member'
} as const;
export type GroupSummaryResponseSettlementPolicyEnum = typeof GroupSummaryResponseSettlementPolicyEnum[keyof typeof GroupSummaryResponseSettlementPolicyEnum];